separate navbar's components
try to render them by breakpoint condition
